chore: 알림 타입 import문 정리
