+++ src/types/notification/index.d.ts
@@ -1,8 +1,1 @@
-export type Notification = {
-  id: number;
-  userId: number;
-  fromId: number;
-  type: number;
-  content: string;
-  isRead: boolean;
-};
+export * from './notificationDto';
