+++ src/types/notification/notificationDto.d.ts
@@ -1,0 +1,8 @@
+export type Notification = {
+  id: number;
+  userId: number;
+  fromId: number;
+  type: number;
+  content: string;
+  isRead: boolean;
+};
